[robustness.improve] use Int64 so that no need use Interlocked.Read, because 64-bit read operations are already atomic. [code.optimize] simplify some code with LINQ.

+++ Plugins.sln/Cube3D/MainWindow.animation.cs
@@ -21,7 +21,7 @@ namespace Cube3D
 {
     public partial class MainWindow
     {
-        private static int _mainWindowRunningAnimationCount;
+        private static long _mainWindowRunningAnimationCount;
 
         private readonly ThicknessAnimation _animationNotifyGrid = new()
         {
